Move IconList to separate component
