+++ eslint-plugin-amsterdam/base.js
@@ -162,7 +162,7 @@ module.exports = {
     'no-extend-native': 'error',
     'no-extra-bind': 'error',
     'no-extra-label': 'error',
-    'no-extra-parens': 'error',
+    'no-extra-parens': 'off',
     'no-extra-semi': 'error',
     'no-fallthrough': 'error',
     'no-floating-decimal': 'error',
